task14

+++ 14-binary_tree_balance.c
@@ -50,7 +50,7 @@ size_t binary_tree_height(const binary_tree_t *tree)
  */
 int binary_tree_balance(const binary_tree_t *tree)
 {
-	size_t le, ri;
+	int le, ri;
 
 	if (tree)
 	{
